requirements.txt
# ============================================
# Core FastAPI Dependencies
# ============================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ============================================
# AI/ML Dependencies
# ============================================
openai==1.3.0
chromadb==0.4.18
sentence-transformers==2.2.2

# ============================================
# FHIR Dependencies
# ============================================
fhirclient==4.1.0
fhir.resources>=7.1.0
httpx==0.25.1

# ============================================
# Database Dependencies
# ============================================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
asyncpg==0.29.0

# ============================================
# Data Generation Dependencies
# (for synthetic data creation)
# ============================================
faker==20.1.0

# ============================================
# Testing Dependencies
# ============================================
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# ============================================
# Utilities
# ============================================
python-dateutil==2.8.2
pytz==2023.3
requests>=2.28.0

# ============================================
# Google Cloud Dependencies
# ============================================
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
